refactor: file structure in ContentPages unified
