+++ frontend/src/components/ContentPages/MyAccountPage/MyAccount.js
@@ -1,10 +1,23 @@
 import React from "react"
+import Navbar from './MyAccountNavbar.js'
+import Header from '../../Header'
+import Footer from '../../Footer'
+import MyAccountContent from "./MyAccountContent.js"
 
 function MyAccount() {
     return (
-        <div className="PageContent">
-            
+        <div>
+        <center>
+        <div className='background'>
+            <Header />
+                <main className="MainContent">
+                    <MyAccountContent />
+                    <Navbar />
+                </main>
+            <Footer />
         </div>
+        </center>
+    </div>
 )
 }
