+++ frontend/src/components/ContentPages/MyAccountPage/MyAccountContent.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+function MyAccountContent() {
+    return (
+        <div className="PageContent">
+            
+        </div>
+)
+}
+
+export default MyAccountContent
