+++ frontend/src/components/ContentPages/MyCrawlsPage/MyCrawls.js
@@ -1,11 +1,24 @@
 import React from "react"
+import Navbar from './MyCrawlsNavbar.js'
+import Header from '../../Header'
+import Footer from '../../Footer'
+import MyCrawlsContent from "./MyCrawlsContent.js"
 
 function MyCrawls() {
     return (
-        <div className="PageContent">
-            
+        <div>
+        <center>
+        <div className='background'>
+            <Header />
+                <main className="MainContent">
+                    <MyCrawlsContent />
+                    <Navbar />
+                </main>
+            <Footer />
         </div>
-)
+        </center>
+    </div>
+    )
 }
 
 export default MyCrawls
