+++ frontend/src/components/ContentPages/MyCrawlsPage/MyCrawlsContent.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+function MyCrawlsContent() {
+    return (
+        <div className="PageContent">
+            
+        </div>
+)
+}
+
+export default MyCrawlsContent
